feat: Set is_primary to 'Y' when creating a new address
This commit modifies the `user.service.ts` file to set the `is_primary` field to 'Y' when creating a new address. This change ensures that the newly created address is marked as the primary address for the user.

+++ src/modules/sql/learning-question-set/learning-question-set.controller.ts
@@ -277,6 +277,22 @@ export class LearningQuestionSetController {
     @Param('id') id: number,
     @Query() query: any,
   ) {
+    const modules = await this.learningModuleService.$db.findOneRecord({
+      options: {
+        where: {
+          question_set_id: +id,
+          active: true,
+        },
+      },
+    });
+    if (modules.data) {
+      return ErrorResponse(res, {
+        error:
+          'Cannot delete this question set as it is currently used in an exam.',
+        message: `Cannot delete this question set as it is currently used in an exam.`,
+      });
+    }
+
     const { error, data } = await this.learningQuestionSetService.delete({
       owner,
       action: 'delete',
